got server running and returning data
