--- controllers/controller.js
@@ -1,6 +1,0 @@
-//import services
-//CRUD operations happen here
-
-module.exports = {
-  
-}
