+++ controllers/login.ctrl.js
@@ -1,0 +1,9 @@
+/*jshint node: true*/
+
+exports.get = function (req, res) {
+    res.send('login get');
+};
+
+exports.post = function (req, res) {
+    res.send('login.post');
+};
